changed calls to callFactories, changed tests to match factoory syntax over constructor syntax, half passing
