--- __tests__/calls/unaryCall.js
@@ -1,8 +1,0 @@
-const UnaryCall = require('../../lib/calls/unaryCall');
-
-const fakeCall = () => {};
-describe('Unit test for Unary Call', () => {
-  test('Unary Call must take a callback', () => {
-    expect(() => new UnaryCall(fakeCall)).toThrow();
-  })
-})
