+++ lib/callFactories/generateDuplex.js
