+++ lib/compose.js
@@ -1,7 +1,7 @@
-const generateUnaryCall = require("./calls/generateUnaryCall");
-const generateClientStreamCall = require("./calls/generateClientStreamCall");
-const generateServerStreamCall = require("./calls/generateServerStreamCall");
-const generateDuplex = require("./calls/generateDuplex");
+const generateUnaryCall = require("./callFactories/generateUnaryCall");
+const generateClientStreamCall = require("./callFactories/generateClientStreamCall");
+const generateServerStreamCall = require("./callFactories/generateServerStreamCall");
+const generateDuplex = require("./callFactories/generateDuplex");
 const getMethodType = require("./utils/getMethodType");
 
 module.exports = function compose(
